+++ jest.config.js
@@ -7,5 +7,13 @@ module.exports = {
   collectCoverageFrom: ['src/**/*.js'],
   coverageDirectory: 'coverage',
   coverageReporters: ['text', 'lcov', 'html'],
+  coverageThreshold: {
+    global: {
+      branches: 50,
+      functions: 50,
+      lines: 70,
+      statements: 70
+    }
+  },
   verbose: true
 };